requirements.txt
awscli==1.11.44
cliff==2.3.0
datakit-core==0.2.0
